+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # [develop](https://github.com/adhearsion/blather/compare/master...develop)
   * Bugfix: Properly sort resources with the same priority but different status
+  * Feature: Implement Blather::Roster#version, which returns version of last processed roster stanza
 
 # [v1.1.4](https://github.com/adhearsion/blather/compare/v1.1.4...v1.1.4) - [2015-06-12](https://rubygems.org/gems/blather/versions/1.1.4)
   * Bugfix: Typo in passing through connection options